Added inheritance & random password generation
Added inheritance to the file manager to allow a file to be written to and allows the previous contents of the file to be overwritten instead of just added to.

Also completed the functionality of the  random password generation so now actually works with the extended character range. Although this implementation may change as it uses a 2D vector to store the Ints that represent the characters and not actually characters as the compiler doesn't support correct type conversion from special characters and provides the incorrect ASCII value